@@ -1,7 +1,7 @@
 {
-  "name": "react-date-range",
-  "version": "1.0.0-beta",
-  "description": "A React component for choosing dates and date ranges.",
+  "name": "@opsrampjoe/date-time-range-picker",
+  "version": "1.0.12",
+  "description": "A React component for choosing date and time ranges.",
   "main": "dist/index.js",
   "scripts": {
     "dev": "NODE_ENV=development & webpack-dev-server --hot --inline",
@@ -22,12 +22,13 @@
   ],
   "contributors": [
     "Burak Can <neoberg@gmail.com> (https://github.com/burakcan)",
-    "Mehmet Kamil Morcay <mehmetmorcay@gmail.com> (https://github.com/mkg0)"
+    "Mehmet Kamil Morcay <mehmetmorcay@gmail.com> (https://github.com/mkg0)",
+    "Joseph R B Taylor <joseph.taylor@opsramp.com> (https://github.com/opsrampjoe)"
   ],
   "license": "MIT",
   "repository": {
     "type": "git",
-    "url": "http://github.com/Adphorus/react-date-range"
+    "url": "https://github.com/opsrampjoe/react-date-range"
   },
   "bugs": {
     "url": "http://github.com/Adphorus/react-date-range/issues"
@@ -35,11 +36,13 @@
   "dependencies": {
     "classnames": "^2.2.1",
     "date-fns": "2.0.0-alpha.7",
+    "postcss": "^7.0.27",
+    "postcss-cli": "^7.1.0",
     "prop-types": "^15.5.10",
     "react-list": "^0.8.8"
   },
   "peerDependencies": {
-    "react": "^0.14 || ^15.0.0-rc || >=15.0"
+    "react": "16.13.0"
   },
   "devDependencies": {
     "autoprefixer": "^7.2.4",
@@ -64,12 +67,12 @@
     "html-webpack-plugin": "^2.30.1",
     "jest": "^22.1.4",
     "normalize.css": "^7.0.0",
-    "postcss-import": "^11.0.0",
-    "postcss-loader": "^2.0.10",
+    "postcss-import": "^12.0.1",
+    "postcss-loader": "^3.0.0",
     "precss": "^2.0.0",
     "prettier": "^1.9.2",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
+    "react": "^16.13.0",
+    "react-dom": "^16.13.0",
     "react-hot-loader": "^3.1.3",
     "style-loader": "^0.19.1",
     "webpack": "^3.10.0",
