@@ -66,15 +66,13 @@ class Calendar extends PureComponent {
         monthHeight: scroll.monthHeight || 220,
         longMonthHeight: longMonthHeight || 260,
         calendarWidth: 'auto',
-        calendarHeight:
-          (scroll.calendarHeight || longMonthHeight || 240) * months,
+        calendarHeight: (scroll.calendarHeight || longMonthHeight || 240) * months,
       };
     }
     return {
       enabled: true,
       monthWidth: scroll.monthWidth || 332,
-      calendarWidth:
-        (scroll.calendarWidth || scroll.monthWidth || 332) * months,
+      calendarWidth: (scroll.calendarWidth || scroll.monthWidth || 332) * months,
       monthHeight: longMonthHeight || 300,
       calendarHeight: longMonthHeight || 300,
     };
@@ -84,11 +82,7 @@ class Calendar extends PureComponent {
       this.setState({ focusedDate: date });
       return;
     }
-    const targetMonthIndex = differenceInCalendarMonths(
-      date,
-      props.minDate,
-      this.dateOptions
-    );
+    const targetMonthIndex = differenceInCalendarMonths(date, props.minDate, this.dateOptions);
     const visibleMonths = this.list.getVisibleRange();
     if (preventUnnecessary && visibleMonths.includes(targetMonthIndex)) return;
     this.list.scrollTo(targetMonthIndex);
@@ -122,7 +116,7 @@ class Calendar extends PureComponent {
       setTimeout(() => this.focusToDate(this.state.focusedDate), 1);
     }
   }
-  componentWillReceiveProps(nextProps) {
+  UNSAFE_componentWillReceiveProps(nextProps) {
     const propMapper = {
       dateRange: 'ranges',
       date: 'date',
@@ -131,9 +125,7 @@ class Calendar extends PureComponent {
     if (this.props.locale !== nextProps.locale) {
       this.dateOptions = { locale: nextProps.locale };
     }
-    if (
-      JSON.stringify(this.props.scroll) !== JSON.stringify(nextProps.scroll)
-    ) {
+    if (JSON.stringify(this.props.scroll) !== JSON.stringify(nextProps.scroll)) {
       this.setState({ scrollArea: this.calcScrollArea(nextProps) });
     }
     if (nextProps[targetProp] !== this.props[targetProp]) {
@@ -154,8 +146,7 @@ class Calendar extends PureComponent {
     onShownDateChange && onShownDateChange(newDate);
   }
   handleRangeFocusChange(rangesIndex, rangeItemIndex) {
-    this.props.onRangeFocusChange &&
-      this.props.onRangeFocusChange([rangesIndex, rangeItemIndex]);
+    this.props.onRangeFocusChange && this.props.onRangeFocusChange([rangesIndex, rangeItemIndex]);
   }
   handleScroll() {
     const { onShownDateChange, minDate } = this.props;
@@ -163,41 +154,24 @@ class Calendar extends PureComponent {
     // prevent scroll jump with wrong visible value
     if (visibleMonths[0] === undefined) return;
     const visibleMonth = addMonths(minDate, visibleMonths[0] || 0);
-    const isFocusedToDifferent = !isSameMonth(
-      visibleMonth,
-      this.state.focusedDate
-    );
+    const isFocusedToDifferent = !isSameMonth(visibleMonth, this.state.focusedDate);
     if (isFocusedToDifferent) {
       this.setState({ focusedDate: visibleMonth });
       onShownDateChange && onShownDateChange(visibleMonth);
     }
   }
   renderMonthAndYear(focusedDate, changeShownDate, props) {
-    const {
-      showMonthArrow,
-      locale,
-      minDate,
-      maxDate,
-      showMonthAndYearPickers,
-    } = props;
-    const upperYearLimit = (
-      maxDate || Calendar.defaultProps.maxDate
-    ).getFullYear();
-    const lowerYearLimit = (
-      minDate || Calendar.defaultProps.minDate
-    ).getFullYear();
+    const { showMonthArrow, locale, minDate, maxDate, showMonthAndYearPickers } = props;
+    const upperYearLimit = (maxDate || Calendar.defaultProps.maxDate).getFullYear();
+    const lowerYearLimit = (minDate || Calendar.defaultProps.minDate).getFullYear();
     const styles = this.styles;
     return (
-      <div
-        onMouseUp={e => e.stopPropagation()}
-        className={styles.monthAndYearWrapper}
-      >
+      <div onMouseUp={e => e.stopPropagation()} className={styles.monthAndYearWrapper}>
         {showMonthArrow ? (
           <button
             type="button"
             className={classnames(styles.nextPrevButton, styles.prevButton)}
-            onClick={() => changeShownDate(-1, 'monthOffset')}
-          >
+            onClick={() => changeShownDate(-1, 'monthOffset')}>
             <i />
           </button>
         ) : null}
@@ -206,8 +180,7 @@ class Calendar extends PureComponent {
             <span className={styles.monthPicker}>
               <select
                 value={focusedDate.getMonth()}
-                onChange={e => changeShownDate(e.target.value, 'setMonth')}
-              >
+                onChange={e => changeShownDate(e.target.value, 'setMonth')}>
                 {locale.localize.months().map((month, i) => (
                   <option key={i} value={i}>
                     {month}
@@ -219,8 +192,7 @@ class Calendar extends PureComponent {
             <span className={styles.yearPicker}>
               <select
                 value={focusedDate.getFullYear()}
-                onChange={e => changeShownDate(e.target.value, 'setYear')}
-              >
+                onChange={e => changeShownDate(e.target.value, 'setYear')}>
                 {new Array(upperYearLimit - lowerYearLimit + 1)
                   .fill(upperYearLimit)
                   .map((val, i) => {
@@ -236,16 +208,14 @@ class Calendar extends PureComponent {
           </span>
         ) : (
           <span className={styles.monthAndYearPickers}>
-            {locale.localize.months()[focusedDate.getMonth()]}{' '}
-            {focusedDate.getFullYear()}
+            {locale.localize.months()[focusedDate.getMonth()]} {focusedDate.getFullYear()}
           </span>
         )}
         {showMonthArrow ? (
           <button
             type="button"
             className={classnames(styles.nextPrevButton, styles.nextButton)}
-            onClick={() => changeShownDate(+1, 'monthOffset')}
-          >
+            onClick={() => changeShownDate(+1, 'monthOffset')}>
             <i />
           </button>
         ) : null}
@@ -274,24 +244,18 @@ class Calendar extends PureComponent {
     return (
       <div className={styles.dateDisplayWrapper}>
         {ranges.map((range, i) => {
-          if (
-            range.showDateDisplay === false ||
-            (range.disabled && !range.showDateDisplay)
-          )
+          if (range.showDateDisplay === false || (range.disabled && !range.showDateDisplay))
             return null;
           return (
             <div
               className={styles.dateDisplay}
               key={i}
-              style={{ color: range.color || defaultColor }}
-            >
+              style={{ color: range.color || defaultColor }}>
               <span
                 className={classnames(styles.dateDisplayItem, {
-                  [styles.dateDisplayItemActive]:
-                    focusedRange[0] === i && focusedRange[1] === 0,
+                  [styles.dateDisplayItemActive]: focusedRange[0] === i && focusedRange[1] === 0,
                 })}
-                onFocus={() => this.handleRangeFocusChange(i, 0)}
-              >
+                onFocus={() => this.handleRangeFocusChange(i, 0)}>
                 <input
                   disabled={range.disabled}
                   readOnly
@@ -302,22 +266,16 @@ class Calendar extends PureComponent {
                   type={showTime ? 'time' : 'hidden'}
                   value={this.formatTimeDisplay(range.startDate)}
                   onChange={evt => {
-                    this.updateTime(
-                      evt.target.value,
-                      range.startDate,
-                      range.endDate,
-                      0
-                    );
+                    this.updateTime(evt.target.value, range.startDate, range.endDate, 0);
                   }}
+                  required={showTime ? true : false}
                 />
               </span>
               <span
                 className={classnames(styles.dateDisplayItem, {
-                  [styles.dateDisplayItemActive]:
-                    focusedRange[0] === i && focusedRange[1] === 1,
+                  [styles.dateDisplayItemActive]: focusedRange[0] === i && focusedRange[1] === 1,
                 })}
-                onFocus={() => this.handleRangeFocusChange(i, 1)}
-              >
+                onFocus={() => this.handleRangeFocusChange(i, 1)}>
                 <input
                   disabled={range.disabled}
                   readOnly
@@ -328,13 +286,9 @@ class Calendar extends PureComponent {
                   type={showTime ? 'time' : 'hidden'}
                   value={this.formatTimeDisplay(range.endDate)}
                   onChange={evt => {
-                    this.updateTime(
-                      evt.target.value,
-                      range.startDate,
-                      range.endDate,
-                      1
-                    );
+                    this.updateTime(evt.target.value, range.startDate, range.endDate, 1);
                   }}
+                  required={showTime ? true : false}
                 />
               </span>
             </div>
@@ -360,12 +314,7 @@ class Calendar extends PureComponent {
   }
 
   onDragSelectionEnd(date) {
-    const {
-      updateRange,
-      displayMode,
-      onChange,
-      dragSelectionEnabled,
-    } = this.props;
+    const { updateRange, displayMode, onChange, dragSelectionEnabled } = this.props;
 
     if (!dragSelectionEnabled) return;
 
@@ -378,10 +327,7 @@ class Calendar extends PureComponent {
       endDate: date,
     };
     if (displayMode !== 'dateRange' || isSameDay(newRange.startDate, date)) {
-      this.setState(
-        { drag: { status: false, range: {} } },
-        () => onChange && onChange(date)
-      );
+      this.setState({ drag: { status: false, range: {} } }, () => onChange && onChange(date));
     } else {
       this.setState({ drag: { status: false, range: {} } }, () => {
         updateRange && updateRange(newRange);
@@ -410,8 +356,7 @@ class Calendar extends PureComponent {
     if (direction === 'horizontal') return scrollArea.monthWidth;
     const monthStep = addMonths(minDate, index);
     const { start, end } = getMonthDisplayRange(monthStep, this.dateOptions);
-    const isLongMonth =
-      differenceInDays(end, start, this.dateOptions) + 1 > 7 * 5;
+    const isLongMonth = differenceInDays(end, start, this.dateOptions) + 1 > 7 * 5;
     return isLongMonth ? scrollArea.longMonthHeight : scrollArea.monthHeight;
   }
   formatDateDisplay(date, defaultText) {
@@ -453,8 +398,7 @@ class Calendar extends PureComponent {
     } = this.props;
     const { scrollArea, focusedDate } = this.state;
     const isVertical = direction === 'vertical';
-    const navigatorRenderer =
-      this.props.navigatorRenderer || this.renderMonthAndYear;
+    const navigatorRenderer = this.props.navigatorRenderer || this.renderMonthAndYear;
 
     const ranges = this.props.ranges.map((range, i) => ({
       ...range,
@@ -462,15 +406,11 @@ class Calendar extends PureComponent {
     }));
     return (
       <div
-        className={classnames(
-          this.styles.calendarWrapper,
-          this.props.className
-        )}
+        className={classnames(this.styles.calendarWrapper, this.props.className)}
         onMouseUp={() => this.setState({ drag: { status: false, range: {} } })}
         onMouseLeave={() => {
           this.setState({ drag: { status: false, range: {} } });
-        }}
-      >
+        }}>
         {showDateDisplay && this.renderDateDisplay()}
         {navigatorRenderer(focusedDate, this.changeShownDate, this.props)}
         {scroll.enabled ? (
@@ -479,17 +419,14 @@ class Calendar extends PureComponent {
             <div
               className={classnames(
                 this.styles.infiniteMonths,
-                isVertical
-                  ? this.styles.monthsVertical
-                  : this.styles.monthsHorizontal
+                isVertical ? this.styles.monthsVertical : this.styles.monthsHorizontal
               )}
               onMouseLeave={() => onPreviewChange && onPreviewChange()}
               style={{
                 width: scrollArea.calendarWidth + 11,
                 height: scrollArea.calendarHeight + 11,
               }}
-              onScroll={this.handleScroll}
-            >
+              onScroll={this.handleScroll}>
               <ReactList
                 length={differenceInCalendarMonths(
                   endOfMonth(maxDate),
@@ -506,9 +443,7 @@ class Calendar extends PureComponent {
                   return (
                     <Month
                       {...this.props}
-                      onPreviewChange={
-                        this.props.onPreviewChange || this.updatePreview
-                      }
+                      onPreviewChange={this.props.onPreviewChange || this.updatePreview}
                       preview={this.props.preview || this.state.preview}
                       ranges={ranges}
                       key={key}
@@ -540,19 +475,14 @@ class Calendar extends PureComponent {
           <div
             className={classnames(
               this.styles.months,
-              isVertical
-                ? this.styles.monthsVertical
-                : this.styles.monthsHorizontal
-            )}
-          >
+              isVertical ? this.styles.monthsVertical : this.styles.monthsHorizontal
+            )}>
             {new Array(this.props.months).fill(null).map((_, i) => {
               const monthStep = addMonths(this.state.focusedDate, i);
               return (
                 <Month
                   {...this.props}
-                  onPreviewChange={
-                    this.props.onPreviewChange || this.updatePreview
-                  }
+                  onPreviewChange={this.props.onPreviewChange || this.updatePreview}
                   preview={this.props.preview || this.state.preview}
                   ranges={ranges}
                   key={i}
@@ -589,14 +519,14 @@ Calendar.defaultProps = {
   showPreview: true,
   displayMode: 'date',
   months: 1,
-  color: '#3d91ff',
+  color: '#424242',
   scroll: {
     enabled: false,
   },
   direction: 'vertical',
   maxDate: addYears(new Date(), 20),
   minDate: addYears(new Date(), -100),
-  rangeColors: ['#3d91ff', '#3ecf8e', '#fed14c'],
+  rangeColors: ['#0077C8', '#3ecf8e', '#fed14c'],
   dragSelectionEnabled: true,
 };
 
